Addressed PR comments
