+++ coulomb-mks-units/src/main/scala/coulomb/mks/package.scala
@@ -69,10 +69,4 @@ package object mks {
   trait Siemens
   implicit val defineUnitSiemens = DerivedUnit[Siemens, (Second %^ 3) %* (Ampere %^ 2) %/ (Kilogram %* (Meter %^ 2))](abbv = "S")
 
-  trait ElectronVolt
-  implicit val defineUnitElectronVolt = {
-    val denom = Rational(10).pow(19)
-    val num = Rational(1602176634L) / Rational(10).pow(9)
-    DerivedUnit[ElectronVolt, Joule](abbv = "eV", coef = num / denom)
-  }
 }
